@@ -19,7 +19,7 @@ end
 def create_user
   create_visitor
   delete_user
-  @user = FactoryGirl.create(:user, email: @visitor[:email])
+  @user = FactoryBot.create(:user, email: @visitor[:email])
   @user.add_role(@visitor[:role])
 end
 
