@@ -1,5 +1,5 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :role do
-    name 'admin'
+    name {'admin'}
   end
 end
