@@ -1,11 +1,11 @@
-# Read about factories at https://github.com/thoughtbot/factory_girl
+# Read about factories at https://github.com/thoughtbot/factory_Bot
 
-FactoryGirl.define do
+FactoryBot.define do
   factory :user do
-    name 'Test User'
-    email 'example@example.com'
-    password 'changeme'
-    password_confirmation 'changeme'
+    name {'Test User'}
+    email {'example@example.com'}
+    password {'changeme'}
+    password_confirmation {'changeme'}
     # required if the Devise Confirmable module is used
     # confirmed_at Time.now
   end
