@@ -22,7 +22,7 @@ module AcademiaRails
     config.generators do |g|
 
       g.test_framework :rspec, fixture: true
-      g.fixture_replacement :factory_girl, dir: 'spec/factories'
+      g.fixture_replacement :factory_Bot, dir: 'spec/factories'
 
 
       g.view_specs false
